requirements.txt
rasa_nlu==0.15.1
openai==0.26.5
spacy==2.3.7
playsound==1.2.2
gtts==2.3.1
google==3.0.0
beautifulsoup4==4.9.3
requests==2.28.2
googlesearch-python==1.1.0
pathlib==1.0.1
openai-whisper==20230124
pyaudio==0.2.13